[storage] downloader_queue file is renamed to downloader_queue_interface
