+++ storage/CMakeLists.txt
@@ -28,7 +28,7 @@ set(
   diff_scheme/diff_types.hpp
   downloader.hpp
   downloader_search_params.hpp
-  downloader_queue.hpp
+  downloader_queue_interface.hpp
   downloader_queue_universal.cpp
   downloader_queue_universal.hpp
   downloading_policy.cpp
